+++ www/client/entries/builder.jsx
@@ -3,4 +3,4 @@ import PageErrorBoundary from "../components/PageErrorBoundary.jsx";
 import Builder from "../features/builder/Builder.jsx";
 import {mountReactRoot} from "../lib/mount-react-root.js";
 
-mountReactRoot({name: "builder", Component: function BuilderPage() { return createElement(PageErrorBoundary, null, createElement(Builder)); }});
+mountReactRoot({name: "builder", Component: function BuilderPage(props) { return createElement(PageErrorBoundary, null, createElement(Builder, props)); }});